Move l.plugin to l.lib.plugin.
