+++ landscape/sysinfo/tests/test_sysinfo.py
@@ -5,7 +5,7 @@ import os
 from twisted.internet.defer import Deferred, succeed, fail
 
 from landscape.lib.compat import StringIO
-from landscape.plugin import PluginRegistry
+from landscape.lib.plugin import PluginRegistry
 from landscape.sysinfo.sysinfo import SysInfoPluginRegistry, format_sysinfo
 from landscape.tests.helpers import LandscapeTest
